Move auth and account-lifecycle routes to auth blueprint
Extracts seven routes from _api.py into the existing auth blueprint
(login, register, change-password, me, logout POST, the two Google
OAuth completion endpoints, and the root login_redirect) and points
login_manager.login_view at the new home. Adds a dedicated
app/serializers/user_serializer.py for the small user_json helper that
several auth routes share.

Also removes the dead /check-username duplicate from _api.py. Both
blueprints registered the same (path, methods) rule, and Flask used
the last-registered view, which was auth.py's. The _api.py copy was
shipped code that never ran. This follows the precedent from commit
830c792 (Dedupe overlapping route registrations). User-visible
behavior of /check-username is unchanged - it was always served by
auth.py's body. The URL surface fixture is regenerated to drop the
duplicate rule entry; no live URLs change.
